requirements.txt
python-telegram-bot
unidecode